player cannot left or right click in the labirynth, player spawn position distant from Lily's position

+++ main/java/mc/mod/prove/eventhandler/PlayerAttackHandler.java
@@ -17,7 +17,7 @@ public class PlayerAttackHandler {
 		if (player != null) {
 			EntityLivingBase entity = event.getEntityLiving();
 
-			int distance = (int) entity.getPositionVector().distanceTo(
+			double distance = entity.getPositionVector().distanceTo(
 					player.getPositionVector());
 
 			if (entity instanceof EntityLilyMob) {
